+++ verticapy/_config/_notebook.py
@@ -1,0 +1,25 @@
+"""
+(c)  Copyright  [2018-2023]  OpenText  or one of its
+affiliates.  Licensed  under  the   Apache  License,
+Version 2.0 (the  "License"); You  may  not use this
+file except in compliance with the License.
+
+You may obtain a copy of the License at:
+http://www.apache.org/licenses/LICENSE-2.0
+
+Unless  required  by applicable  law or  agreed to in
+writing, software  distributed  under the  License is
+distributed on an  "AS IS" BASIS,  WITHOUT WARRANTIES
+OR CONDITIONS OF ANY KIND, either express or implied.
+See the  License for the specific  language governing
+permissions and limitations under the License.
+"""
+ISNOTEBOOK = False
+try:
+    import IPython
+
+    shell = get_ipython().__class__.__name__
+    if shell == "ZMQInteractiveShell":
+        ISNOTEBOOK = True  # Jupyter notebook or qtconsole
+except:
+    pass
