+++ verticapy/machine_learning/vertica/delphi.py
@@ -32,6 +32,7 @@ from verticapy.utils._decorators import (
 from verticapy import vDataFrame
 from verticapy.utilities import *
 from verticapy.utils._toolbox import *
+from verticapy.utils._gen import gen_tmp_name
 from verticapy.errors import *
 from verticapy.learn.ensemble import *
 from verticapy.learn.naive_bayes import *
