+++ verticapy/machine_learning/vertica/ensemble.py
@@ -33,6 +33,7 @@ from verticapy.learn.metrics import *
 from verticapy.plotting._matplotlib import *
 from verticapy.utilities import *
 from verticapy.utils._toolbox import *
+from verticapy.utils._gen import gen_name
 from verticapy.sql.read import _executeSQL
 from verticapy import vDataFrame
 from verticapy.errors import *
