+++ verticapy/machine_learning/vertica/preprocessing.py
@@ -30,6 +30,7 @@ from verticapy.utils._decorators import (
 )
 from verticapy.utilities import *
 from verticapy.utils._toolbox import *
+from verticapy.utils._gen import gen_tmp_name
 from verticapy.sql.read import _executeSQL
 from verticapy import vDataFrame
 from verticapy.learn.vmodel import *
