+++ verticapy/machine_learning/vertica/vmodel.py
@@ -35,6 +35,7 @@ from verticapy.plotting._matplotlib import *
 from verticapy.learn.model_selection import *
 from verticapy.utilities import *
 from verticapy.utils._toolbox import *
+from verticapy.utils._gen import gen_name, gen_tmp_name
 from verticapy.sql.read import _executeSQL
 from verticapy.errors import *
 import verticapy.learn.metrics as mt
