+++ verticapy/plotting/_matplotlib/animated.py
@@ -27,7 +27,8 @@ import numpy as np
 
 # VerticaPy Modules
 from verticapy.utilities import *
-from verticapy.utils._toolbox import updated_dict, isnotebook
+from verticapy.utils._toolbox import updated_dict
+from verticapy._config._notebook import ISNOTEBOOK
 from verticapy.sql.read import _executeSQL
 from verticapy.errors import ParameterError
 from verticapy.plotting._colors import gen_cmap, gen_colors
@@ -155,7 +156,7 @@ def animated_bar(
             current_ts, ts_idx = elem, idx
     if not (ax):
         fig, ax = plt.subplots()
-        if isnotebook():
+        if ISNOTEBOOK:
             if pie:
                 fig.set_size_inches(11, min(limit_over, 600))
             else:
@@ -299,7 +300,7 @@ def animated_bar(
         blit=False,
         repeat=repeat,
     )
-    if isnotebook() and return_html:
+    if ISNOTEBOOK and return_html:
         anim = myAnimation.to_jshtml()
         plt.close("all")
         return HTML(anim)
@@ -376,7 +377,7 @@ def animated_bubble_plot(
     count = vdf.shape()[0]
     if not (ax):
         fig, ax = plt.subplots()
-        if isnotebook():
+        if ISNOTEBOOK:
             fig.set_size_inches(12, 8)
         ax.grid()
         ax.set_axisbelow(True)
@@ -568,7 +569,7 @@ def animated_bubble_plot(
         blit=False,
         repeat=repeat,
     )
-    if isnotebook() and return_html:
+    if ISNOTEBOOK and return_html:
         anim = myAnimation.to_jshtml()
         plt.close("all")
         return HTML(anim)
@@ -638,7 +639,7 @@ def animated_ts_plot(
     alpha = 0.3
     if not (ax):
         fig, ax = plt.subplots()
-        if isnotebook():
+        if ISNOTEBOOK:
             fig.set_size_inches(8, 6)
         ax.grid(axis="y")
         ax.set_axisbelow(True)
@@ -694,7 +695,7 @@ def animated_ts_plot(
         blit=False,
         repeat=repeat,
     )
-    if isnotebook() and return_html:
+    if ISNOTEBOOK and return_html:
         anim = myAnimation.to_jshtml()
         plt.close("all")
         return HTML(anim)
