+++ verticapy/plotting/_matplotlib/mlplot.py
@@ -36,6 +36,7 @@ from verticapy.utils._decorators import (
 )
 from verticapy.utilities import *
 from verticapy.utils._toolbox import *
+from verticapy._config._notebook import ISNOTEBOOK
 from verticapy.sql.read import _executeSQL
 from verticapy.errors import *
 from verticapy.plotting._colors import gen_colors
@@ -85,7 +86,7 @@ def logit_plot(
         )
         if not (ax):
             fig, ax = plt.subplots()
-            if isnotebook():
+            if ISNOTEBOOK:
                 fig.set_size_inches(8, 6)
             ax.set_axisbelow(True)
             ax.grid()
@@ -179,7 +180,7 @@ def logit_plot(
             )
         )
         if not (ax):
-            if isnotebook():
+            if ISNOTEBOOK:
                 plt.figure(figsize=(8, 6))
             ax = plt.axes(projection="3d")
         ax.plot_surface(
@@ -283,7 +284,7 @@ def lof_plot(
         column2 = [0] * len(column1)
         if not (ax):
             fig, ax = plt.subplots()
-            if isnotebook():
+            if ISNOTEBOOK:
                 fig.set_size_inches(8, 2)
             ax.set_axisbelow(True)
             ax.grid()
@@ -322,7 +323,7 @@ def lof_plot(
         )
         if not (ax):
             fig, ax = plt.subplots()
-            if isnotebook():
+            if ISNOTEBOOK:
                 fig.set_size_inches(8, 6)
             ax.set_axisbelow(True)
             ax.grid()
@@ -363,7 +364,7 @@ def lof_plot(
             [float(item[3]) for item in query_result],
         )
         if not (ax):
-            if isnotebook():
+            if ISNOTEBOOK:
                 plt.figure(figsize=(8, 6))
             ax = plt.axes(projection="3d")
         ax.set_xlabel(columns[0])
@@ -407,7 +408,7 @@ def plot_importance(
     )
     if not (ax):
         fig, ax = plt.subplots()
-        if isnotebook():
+        if ISNOTEBOOK:
             fig.set_size_inches(12, int(len(importances) / 2) + 1)
         ax.set_axisbelow(True)
         ax.grid()
@@ -449,7 +450,7 @@ def plot_stepwise_ml(
     colors = gen_colors()
     if not (ax):
         fig, ax = plt.subplots()
-        if isnotebook():
+        if ISNOTEBOOK:
             fig.set_size_inches(8, 6)
         ax.grid(axis="y")
         ax.set_axisbelow(True)
@@ -561,7 +562,7 @@ def plot_bubble_ml(
     colors = gen_colors()
     if not (ax):
         fig, ax = plt.subplots()
-        if isnotebook():
+        if ISNOTEBOOK:
             fig.set_size_inches(8, 6)
         ax.grid(axis="y")
         ax.set_axisbelow(True)
@@ -716,7 +717,7 @@ def plot_pca_circle(
     circle1 = plt.Circle((0, 0), 1, edgecolor=colors[0], facecolor="none")
     if not (ax):
         fig, ax = plt.subplots()
-        if isnotebook():
+        if ISNOTEBOOK:
             fig.set_size_inches(6, 6)
         ax.set_axisbelow(True)
     n = len(x)
@@ -760,7 +761,7 @@ def plot_var(
         colors[0] = style_kwds["color"]
     if not (ax):
         fig, ax = plt.subplots()
-        if isnotebook():
+        if ISNOTEBOOK:
             fig.set_size_inches(6, 6)
         ax.set_axisbelow(True)
         ax.grid()
@@ -836,7 +837,7 @@ def regression_plot(
         )
         if not (ax):
             fig, ax = plt.subplots()
-            if isnotebook():
+            if ISNOTEBOOK:
                 fig.set_size_inches(8, 6)
             ax.set_axisbelow(True)
             ax.grid()
@@ -891,7 +892,7 @@ def regression_plot(
         X_reg, Y_reg = np.meshgrid(X_reg, Y_reg)
         Z_reg = coefficients[0] + coefficients[1] * X_reg + coefficients[2] * Y_reg
         if not (ax):
-            if isnotebook():
+            if ISNOTEBOOK:
                 plt.figure(figsize=(8, 6))
             ax = plt.axes(projection="3d")
         ax.plot_surface(
@@ -934,7 +935,7 @@ def regression_tree_plot(
     )
     if not (ax):
         fig, ax = plt.subplots()
-        if isnotebook():
+        if ISNOTEBOOK:
             fig.set_size_inches(8, 6)
         ax.set_axisbelow(True)
         ax.grid()
@@ -1001,7 +1002,7 @@ def svm_classifier_plot(
         )
         if not (ax):
             fig, ax = plt.subplots()
-            if isnotebook():
+            if ISNOTEBOOK:
                 fig.set_size_inches(8, 6)
             ax.set_axisbelow(True)
             ax.grid()
@@ -1051,7 +1052,7 @@ def svm_classifier_plot(
         )
         if not (ax):
             fig, ax = plt.subplots()
-            if isnotebook():
+            if ISNOTEBOOK:
                 fig.set_size_inches(8, 6)
             ax.set_axisbelow(True)
             ax.grid()
@@ -1131,7 +1132,7 @@ def svm_classifier_plot(
         X_svm, Y_svm = np.meshgrid(X_svm, Y_svm)
         Z_svm = coefficients[0] + coefficients[1] * X_svm + coefficients[2] * Y_svm
         if not (ax):
-            if isnotebook():
+            if ISNOTEBOOK:
                 plt.figure(figsize=(8, 6))
             ax = plt.axes(projection="3d")
         ax.plot_surface(
