+++ verticapy/sdk/vertica/dblink.py
@@ -41,7 +41,7 @@ def get_dblink_fun(query: str, symbol: str = "$"):
 
 
 def replace_external_queries_in_query(query: str):
-    from verticapy.utils._toolbox import gen_tmp_name
+    from verticapy.utils._gen import gen_tmp_name
     from verticapy.sql.read import _executeSQL
 
     sql_keyword = (
