+++ verticapy/sql/parsers/pandas.py
@@ -22,6 +22,7 @@ import os, csv
 import verticapy as vp
 from verticapy.utils._decorators import save_verticapy_logs
 from verticapy.utils._toolbox import *
+from verticapy.utils._gen import gen_tmp_name
 from verticapy.sql.read import _executeSQL
 from verticapy.errors import ParameterError
 from verticapy.sql.parsers.csv import read_csv
