+++ verticapy/stats/math/math.py
@@ -29,7 +29,7 @@ from verticapy.utils._toolbox import *
 from verticapy.stats.tools import *
 from verticapy.core.str_sql import str_sql
 from verticapy.sql._utils._format import format_magic, clean_query
-from verticapy.utils._map import python_to_dtype_category
+from verticapy.utils._cast import python_to_dtype_category
 
 #
 # Global Variables
